pset7/numb3rs/test_numb3rs.py
from numb3rs import validate  # Import the validate function from numb3rs module

# Test the ip ranges
def test_ipranges():
    # Valid IPv4 addresses
    assert validate("255.255.255.255")
    assert validate("192.168.1.1")
    assert validate("192.168.1.255")
    assert validate("0.0.0.0")

    # Invalid IPv4 addresses
    assert not validate("192.168.1")         # Missing octet
    assert not validate("192.168.1.256")     # Octet value exceeds 255
    assert not validate("192.168.1.255.1")   # Too many octets
    assert not validate("512.512.512.512")   # Octet value exceeds 255
    assert not validate("512.512.512.512.512")  # Too many octets
    assert not validate("cat")               # Non-numeric input

"""
This set of test cases covers a variety of scenarios, including valid and invalid IPv4 addresses, different octet values, and non-numeric inputs. 
These tests will help ensure that your validate function behaves as expected and handles various cases correctly.
"""